Ensuring newly written files can be read by the hive connector by introduction of a refresh frequency configuration option.

+++ kafka-connect-hive/src/test/scala/com/landoop/streamreactor/connect/hive/source/HiveSourceTest.scala
@@ -7,11 +7,14 @@ import com.landoop.streamreactor.connect.hive._
 import com.landoop.streamreactor.connect.hive.sink.HiveSink
 import com.landoop.streamreactor.connect.hive.sink.config.{HiveSinkConfig, TableOptions}
 import com.landoop.streamreactor.connect.hive.source.config.{HiveSourceConfig, ProjectionField, SourceTableOptions}
-import com.landoop.streamreactor.connect.hive.source.offset.{HiveSourceOffsetStorageReader, MockOffsetStorageReader}
+import com.landoop.streamreactor.connect.hive.source.offset.{HiveSourceInitOffsetStorageReader, MockOffsetStorageReader}
 import com.typesafe.scalalogging.StrictLogging
 import org.apache.hadoop.fs.Path
 import org.apache.hadoop.hive.metastore.api.Database
 import org.apache.kafka.connect.data.{SchemaBuilder, Struct}
+import org.apache.kafka.connect.source.SourceTaskContext
+import org.apache.kafka.connect.storage.OffsetStorageReader
+import org.mockito.MockitoSugar.{mock, when}
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 
@@ -40,10 +43,20 @@ class HiveSourceTest extends AnyWordSpec with Matchers with HiveTestConfig with 
     .field("salary", SchemaBuilder.float64().optional().build())
     .build()
 
-  def populateEmployees(table: String, partitions: Seq[PartitionField] = Nil): Unit = {
-    Try {
+  def populateEmployees(table: String, partitions: Seq[PartitionField] = Nil, dropTableFirst: Boolean = true, offsetAdd: Int = 0): Unit = {
+    if (dropTableFirst) Try {
       client.dropTable(dbname, table, true, true)
     }
+
+    val sinkConfig = HiveSinkConfig(DatabaseName(dbname), tableOptions = Set(
+      TableOptions(TableName(table), Topic("mytopic"), dropTableFirst, dropTableFirst, partitions = partitions)
+    ),
+      kerberos = None,
+      hadoopConfiguration = HadoopConfiguration.Empty
+    )
+
+    val sink = HiveSink.from(TableName(table), sinkConfig)
+
     val users = List(
       new Struct(schema).put("name", "sam").put("title", "mr").put("salary", 100.43),
       new Struct(schema).put("name", "laura").put("title", "ms").put("salary", 429.06),
@@ -52,20 +65,12 @@ class HiveSourceTest extends AnyWordSpec with Matchers with HiveTestConfig with 
       new Struct(schema).put("name", "ant").put("title", "mr").put("salary", 629.06),
       new Struct(schema).put("name", "tom").put("title", "miss").put("salary", 395.44)
     )
-
-    val sinkConfig = HiveSinkConfig(DatabaseName(dbname), tableOptions = Set(
-      TableOptions(TableName(table), Topic("mytopic"), true, true, partitions = partitions)
-    ),
-      kerberos = None,
-      hadoopConfiguration = HadoopConfiguration.Empty
-    )
-
-    val sink = HiveSink.from(TableName(table), sinkConfig)
     users.zipWithIndex.foreach { case (user, k) =>
-      sink.write(user, TopicPartitionOffset(Topic("mytopic"), 1, Offset(k)))
+      sink.write(user, TopicPartitionOffset(Topic("mytopic"), 1, Offset(k + offsetAdd)))
     }
     sink.close()
   }
+
 
   "hive source" should {
 
@@ -74,7 +79,7 @@ class HiveSourceTest extends AnyWordSpec with Matchers with HiveTestConfig with 
       val table = "employees"
       populateEmployees(table)
 
-      val reader = new HiveSourceOffsetStorageReader(new MockOffsetStorageReader(Map.empty))
+      val reader = new HiveSourceInitOffsetStorageReader(new MockOffsetStorageReader(Map.empty))
       val sourceConfig = HiveSourceConfig(DatabaseName(dbname), tableOptions = Set(
         SourceTableOptions(TableName(table), Topic("mytopic"))
       ), kerberos = None,
@@ -89,7 +94,7 @@ class HiveSourceTest extends AnyWordSpec with Matchers with HiveTestConfig with 
       val table = "employees"
       populateEmployees(table)
 
-      val reader = new HiveSourceOffsetStorageReader(new MockOffsetStorageReader(Map.empty))
+      val reader = new HiveSourceInitOffsetStorageReader(new MockOffsetStorageReader(Map.empty))
       val sourceConfig = HiveSourceConfig(DatabaseName(dbname), tableOptions = Set(
         SourceTableOptions(TableName(table), Topic("mytopic"), projection = Some(NonEmptyList.of(ProjectionField("title", "title"), ProjectionField("name", "name"))))
       ), kerberos = None,
@@ -106,7 +111,7 @@ class HiveSourceTest extends AnyWordSpec with Matchers with HiveTestConfig with 
       val table = "employees"
       populateEmployees(table)
 
-      val reader = new HiveSourceOffsetStorageReader(new MockOffsetStorageReader(Map.empty))
+      val reader = new HiveSourceInitOffsetStorageReader(new MockOffsetStorageReader(Map.empty))
       val sourceConfig = HiveSourceConfig(DatabaseName(dbname), tableOptions = Set(
         SourceTableOptions(TableName(table), Topic("mytopic"), projection = Some(NonEmptyList.of(ProjectionField("title", "salutation"), ProjectionField("name", "name"))))
       ), kerberos = None,
@@ -123,7 +128,7 @@ class HiveSourceTest extends AnyWordSpec with Matchers with HiveTestConfig with 
       val table = "employees_partitioned"
       populateEmployees(table, partitions = Seq(PartitionField("title")))
 
-      val reader = new HiveSourceOffsetStorageReader(new MockOffsetStorageReader(Map.empty))
+      val reader = new HiveSourceInitOffsetStorageReader(new MockOffsetStorageReader(Map.empty))
       val sourceConfig = HiveSourceConfig(DatabaseName(dbname), tableOptions = Set(
         SourceTableOptions(TableName(table), Topic("mytopic"))
       ), kerberos = None,
@@ -138,7 +143,7 @@ class HiveSourceTest extends AnyWordSpec with Matchers with HiveTestConfig with 
       val table = "employees_partitioned"
       populateEmployees(table, partitions = Seq(PartitionField("title")))
 
-      val reader = new HiveSourceOffsetStorageReader(new MockOffsetStorageReader(Map.empty))
+      val reader = new HiveSourceInitOffsetStorageReader(new MockOffsetStorageReader(Map.empty))
       val sourceConfig = HiveSourceConfig(DatabaseName(dbname), tableOptions = Set(
         SourceTableOptions(TableName(table), Topic("mytopic"), projection = Some(NonEmptyList.of(ProjectionField("title", "title"), ProjectionField("name", "name"))))
       ), kerberos = None,
@@ -152,7 +157,7 @@ class HiveSourceTest extends AnyWordSpec with Matchers with HiveTestConfig with 
       val table = "employees"
       populateEmployees(table)
 
-      val reader = new HiveSourceOffsetStorageReader(new MockOffsetStorageReader(Map.empty))
+      val reader = new HiveSourceInitOffsetStorageReader(new MockOffsetStorageReader(Map.empty))
       val sourceConfig = HiveSourceConfig(DatabaseName(dbname), tableOptions = Set(
         SourceTableOptions(TableName(table), Topic("mytopic"), limit = 3)
       ), kerberos = None,
@@ -167,7 +172,7 @@ class HiveSourceTest extends AnyWordSpec with Matchers with HiveTestConfig with 
       val table = "employees"
       populateEmployees(table)
 
-      val reader = new HiveSourceOffsetStorageReader(new MockOffsetStorageReader(Map.empty))
+      val reader = new HiveSourceInitOffsetStorageReader(new MockOffsetStorageReader(Map.empty))
       val sourceConfig = HiveSourceConfig(DatabaseName(dbname), tableOptions = Set(
         SourceTableOptions(TableName(table), Topic("mytopic"))
       ), kerberos = None,
@@ -199,7 +204,7 @@ class HiveSourceTest extends AnyWordSpec with Matchers with HiveTestConfig with 
         new Path(s"hdfs://localhost:8020/user/hive/warehouse/$dbname/$table/streamreactor_mytopic_1_5")
       )
       val sourceOffset = SourceOffset(2)
-      val reader = new HiveSourceOffsetStorageReader(new MockOffsetStorageReader(Map(sourcePartition -> sourceOffset)))
+      val reader = new HiveSourceInitOffsetStorageReader(new MockOffsetStorageReader(Map(sourcePartition -> sourceOffset)))
 
       val sourceConfig = HiveSourceConfig(DatabaseName(dbname), tableOptions = Set(
         SourceTableOptions(TableName(table), Topic("mytopic"))
@@ -221,7 +226,7 @@ class HiveSourceTest extends AnyWordSpec with Matchers with HiveTestConfig with 
         new Path(s"hdfs://localhost:8020/user/hive/warehouse/$dbname/$table/streamreactor_mytopic_1_5")
       )
       val sourceOffset = SourceOffset(2)
-      val reader = new HiveSourceOffsetStorageReader(new MockOffsetStorageReader(Map(sourcePartition -> sourceOffset)))
+      val reader = new HiveSourceInitOffsetStorageReader(new MockOffsetStorageReader(Map(sourcePartition -> sourceOffset)))
 
       val sourceConfig = HiveSourceConfig(DatabaseName(dbname), tableOptions = Set(
         SourceTableOptions(TableName(table), Topic("mytopic"), limit = 2)
@@ -243,7 +248,7 @@ class HiveSourceTest extends AnyWordSpec with Matchers with HiveTestConfig with 
         new Path(s"hdfs://localhost:8020/user/hive/warehouse/$dbname/$table/streamreactor_mytopic_1_5")
       )
       val sourceOffset = SourceOffset(44)
-      val reader = new HiveSourceOffsetStorageReader(new MockOffsetStorageReader(Map(sourcePartition -> sourceOffset)))
+      val reader = new HiveSourceInitOffsetStorageReader(new MockOffsetStorageReader(Map(sourcePartition -> sourceOffset)))
 
       val sourceConfig = HiveSourceConfig(DatabaseName(dbname), tableOptions = Set(
         SourceTableOptions(TableName(table), Topic("mytopic"), limit = 2)
@@ -251,6 +256,45 @@ class HiveSourceTest extends AnyWordSpec with Matchers with HiveTestConfig with 
         hadoopConfiguration = HadoopConfiguration.Empty)
       val source = new HiveSource(DatabaseName(dbname), TableName(table), Topic("mytopic"), reader, sourceConfig)
       source.toList.isEmpty shouldBe true
+    }
+
+    "discover partitions after the source has started" in {
+
+      val table = "employees"
+      val tblObject = client.getTable(dbname, table)
+
+      client.dropTable(dbname, table, true, true)
+      client.createTable(tblObject)
+
+      val props = Map(
+              "connect.hive.database.name" -> dbname,
+              "connect.hive.metastore" -> "thrift",
+              "connect.hive.metastore.uris" -> "thrift://namenode:9083",
+              "connect.hive.fs.defaultFS" -> "hdfs://namenode:8020",
+              "connect.hive.kcql" -> s"insert into mytopic select name, title, salary from $table",
+              "connect.hive.refresh.frequency" -> "1",
+      )
+
+      val sourceTaskContext: SourceTaskContext = mock[SourceTaskContext]
+      val reader = mock[OffsetStorageReader]
+      when(sourceTaskContext.offsetStorageReader()).thenReturn(reader)
+
+      populateEmployees(table)
+
+      val sourceTask = new HiveSourceTask()
+      sourceTask.initialize(sourceTaskContext)
+      sourceTask.start(props.asJava)
+
+      val res1 = sourceTask.poll()
+      res1.size shouldBe (6)
+
+      populateEmployees(table, dropTableFirst = false, offsetAdd = 6)
+      populateEmployees(table, dropTableFirst = false, offsetAdd = 12)
+
+      Thread.sleep(1200)
+      val res2 = sourceTask.poll()
+      res2.size shouldBe (12)
+
     }
   }
 }
